Move TimeoutManager into its own module
