+++ refbox/ui_test.py
@@ -1,5 +1,5 @@
 from . import ui
-from .gamemanager import GameManager, GameState, TimeoutState
+from .gamemanager import GameManager
 from .noiomanager import IOManager
 
 def test_game_config_parser():
@@ -51,52 +51,3 @@ def test_edit_time():
     nv.mgr.setGameClock(2)
     nv.edit_time()
     assert nv.mgr.gameClock() == 2
-
-
-def test_ref_timeout_clicked():
-    mgr = GameManager()
-    assert mgr.gameClockRunning() is False
-    assert mgr.gameState() == GameState.first_half
-    assert mgr.timeoutState() == TimeoutState.none
-
-    timeout_mgr = ui.TimeoutManager()
-    assert timeout_mgr.text.get() == "START"
-
-    # Test when button says START.
-    timeout_mgr.click(mgr, 0)
-    assert mgr.gameClockRunning() is True
-    assert mgr.gameState() == GameState.first_half
-    assert mgr.timeoutState() == TimeoutState.none
-    assert timeout_mgr.text.get() == "TIMEOUT"
-
-    # Test when button says TIMEOUT.
-    timeout_mgr.click(mgr, 0)
-    assert mgr.gameClockRunning() is False
-    assert mgr.gameState() == GameState.first_half
-    assert mgr.timeoutState() == TimeoutState.ref
-    assert timeout_mgr.text.get() == "RESUME"
-
-    # Test when button says RESUME.
-    timeout_mgr.click(mgr, 0)
-    assert mgr.gameClockRunning() is True
-    assert mgr.gameState() == GameState.first_half
-    assert mgr.timeoutState() == TimeoutState.none
-    assert timeout_mgr.text.get() == "TIMEOUT"
-
-    # Jump to Game Over mode.
-    mgr.setBlackScore(1)
-    mgr.setWhiteScore(1)
-    timeout_mgr.set_game_over(mgr)
-    assert mgr.blackScore() == 1
-    assert mgr.whiteScore() == 1
-    assert mgr.gameClockRunning() is False
-    assert mgr.gameState() == GameState.game_over
-
-    # Test when button says RESET.
-    timeout_mgr.click(mgr, 0)
-    assert mgr.blackScore() == 0
-    assert mgr.whiteScore() == 0
-    assert mgr.gameClockRunning() is False
-    assert mgr.gameState() == GameState.first_half
-    assert mgr.timeoutState() == TimeoutState.none
-    assert timeout_mgr.text.get() == "START"
